backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0
python-dotenv==1.0.0

# Include the existing karaoke-gen package as dependency
# This will install from the parent directory
-e ../

